Struct azure_functions::bindings::TimerInfo

pub struct TimerInfo {
    pub schedule_status: Option<ScheduleStatus>,
    pub is_past_due: bool,
}

Represents the timer information from a timer trigger binding.

Examples

A function that runs every 5 minutes:

use azure_functions::bindings::TimerInfo;
use azure_functions::func;

#[func]
#[binding(name = "info", schedule = "0 */5 * * * *")]
pub fn timer(info: &TimerInfo) {
    info!("Rust Azure function ran!");
}

Fields
schedule_status: Option<ScheduleStatus>

The schedule status for the timer. If schedule monitoring is not enabled for the timer, this field will be None.

is_past_due: bool

Determines if the timer invocation is due to a missed schedule occurrence.