Enum azure_functions::http::Body

pub enum Body<'a> {
    Empty,
    String(Cow<'a, str>),
    Json(Cow<'a, str>),
    Bytes(Cow<'a, [u8]>),
}

Represents the body of a HTTP request or response.

Variants
Empty

Represents an empty body.

String(Cow<'a, str>)

Represents a string body with a default content type of text/plain.

Json(Cow<'a, str>)

Represents a JSON body with a default content type of application/json.

Bytes(Cow<'a, [u8]>)

Represents a body from a slice of bytes with a default content type of application/octet-stream.

Methods

impl<'a> Body<'a>

pub fn default_content_type(&self) -> &str

Gets the default content type for a body.

Returns application/json for Body::Json.

Returns application/octet-stream for Body::Bytes.

Returns text/plain for all other Body values.

Examples

use azure_functions::http::Body;

let body: Body = [1u8, 2u8, 3u8][..].into();

assert_eq!(body.default_content_type(), "application/octet-stream");